Move components with authz to separate folder in auth module and unify code of them
