+++ src/modules/auth/components/UncontrolledSwitchWithAuthz.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { UncontrolledSwitch } from 'asab-webui';
-import { useTranslation } from 'react-i18next';
+import { authz } from './authz';
 
 /*
 	UncontrolledSwitchWithAuthz component emulates the functionality of asab-webui <UncontrolledSwitch/> component
@@ -54,28 +54,12 @@ import { useTranslation } from 'react-i18next';
 */
 
 export function UncontrolledSwitchWithAuthz(props) {
-	const { t, i18n } = useTranslation();
 	let childProps = {...props}; // Create a new child component to have option to remove props
-	let disabled = childProps.resources ? childProps.resources.indexOf(childProps.resource) === -1 : true;
-	// If user is superuser, then button is enabled
-	if (childProps.resources.indexOf('authz:superuser') !== -1) {
-		disabled = false;
-	}
-	// If defined, hide the disabled button
-	let hide = childProps.hideOnUnauthorizedAccess ? true : false;
-	// Remove hideOnUnauthorized element from props to avoid react warnings
-	if (childProps.hideOnUnauthorizedAccess) {
-		delete childProps["hideOnUnauthorizedAccess"];
-	}
-	let title = childProps.title;
-	// Check on title eventually passed in the props
-	if (disabled) {
-		 title = t("You do not have access rights to perform this action");
-	}
-	// Check on disabled eventually passed in the props
-	if (childProps.disabled && disabled == false) {
-		disabled = childProps.disabled;
-	}
+	let authzObj = authz(childProps);
+
+	const disabled = authzObj.disabled;
+	const hide = authzObj.hide;
+	const title = authzObj.title;
 
 	return (
 		hide && disabled ? null :
